MYR-1974: removed duplicate function

+++ src/utils/upload.ts
@@ -6,68 +6,45 @@ import sharp from 'sharp';
 import fs, {existsSync} from 'fs';
 import path from 'path';
 import os from 'os';
+import {UploadType} from '../enums';
 
-export async function upload(data: AnyObject, name: string) {
-  const assetPath = path.join(__dirname, `../../seed-data/assets`);
-
-  if (!existsSync(assetPath)) return '';
-
-  const files = fs.readdirSync(assetPath);
-  const file = files.find(e => {
-    const [imageFile] = e.split(/\.(jpeg|jpg|gif|png)$/);
-    if (imageFile === data.imageFileName) return true;
-    return false;
-  });
-
-  if (!file) return '';
+export async function upload(
+  type: UploadType,
+  targetDir: string,
+  filePath: string,
+) {
+  if (!filePath) return '';
 
   const bucket = config.FIREBASE_STORAGE_BUCKET
     ? firebaseAdmin.storage().bucket()
     : undefined;
 
-  const baseName = uuid();
   const tmpDir = os.tmpdir();
-  const filePath = `${assetPath}/${file}`;
-  const format = 'jpg';
-  const mutations = [
-    {
-      type: 'thumbnail',
-      suffix: '_thumbnail',
-      width: 200,
-    },
-    {
-      type: 'small',
-      suffix: '_small',
-      width: 400,
-    },
-    {
-      type: 'medium',
-      suffix: '_medium',
-      width: 600,
-    },
-    {
-      type: 'origin',
-      suffix: '',
-      width: 0,
-    },
-  ];
+  const baseName = path.parse(filePath).name;
+  const extension = path.parse(filePath).ext;
+  const {format, mutations} = getMutations(UploadType.IMAGE);
 
   let result = '';
   for (const mutation of mutations) {
-    const formattedFilePath = `${tmpDir}/${baseName}${mutation.suffix}_formatted.${format}`;
-    const uploadFilePath = `${data.imagePath}/${name}/${baseName}${mutation.suffix}.${format}`;
+    let formattedFilePath = `${tmpDir}/${baseName}${extension}`;
+    let uploadFilePath = `${targetDir}/${baseName}${extension}`;
 
-    if (mutation.type === 'origin') {
-      await sharp(filePath)
-        .withMetadata()
-        .toFormat('jpg')
-        .toFile(formattedFilePath);
-    } else {
-      await sharp(filePath)
-        .resize({width: mutation.width})
-        .withMetadata()
-        .toFormat('jpg')
-        .toFile(formattedFilePath);
+    if (type === UploadType.IMAGE) {
+      formattedFilePath = `${tmpDir}/${baseName}${mutation.suffix}_formatted.${format}`;
+      uploadFilePath = `${targetDir}/${baseName}${mutation.suffix}.${format}`;
+
+      if (mutation.type === 'origin') {
+        await sharp(filePath)
+          .withMetadata()
+          .toFormat('jpg')
+          .toFile(formattedFilePath);
+      } else {
+        await sharp(filePath)
+          .resize({width: mutation.width})
+          .withMetadata()
+          .toFormat('jpg')
+          .toFile(formattedFilePath);
+      }
     }
 
     if (bucket) {
@@ -80,27 +57,94 @@ export async function upload(data: AnyObject, name: string) {
       result = imageFile.publicUrl();
     } else {
       if (!config.STORAGE_URL) {
+        fs.unlinkSync(filePath);
         fs.unlinkSync(formattedFilePath);
-        throw new Error('Storage Not Found');
+        throw new Error('Storage not found');
       }
 
-      const folderPath = `../../storages/${data.imagePath}/${name}`;
-      const storagePath = path.join(__dirname, `${folderPath}`);
-
-      if (!fs.existsSync(storagePath)) {
-        fs.mkdirSync(storagePath, {recursive: true});
+      const folderPath = '../../storages';
+      const tmpSubFolderPath = `${folderPath}/${targetDir}`;
+      const tmpUpdatedFilePath = `${folderPath}/${uploadFilePath}`;
+      const subfolderPath = path.join(__dirname, tmpSubFolderPath);
+      const updatedFilePath = path.join(__dirname, tmpUpdatedFilePath);
+      if (!fs.existsSync(subfolderPath)) {
+        fs.mkdirSync(subfolderPath, {recursive: true});
       }
 
-      fs.copyFileSync(
-        formattedFilePath,
-        `${storagePath}/${baseName}${mutation.suffix}.${format}`,
-      );
+      fs.copyFileSync(formattedFilePath, updatedFilePath);
 
       result = `${config.STORAGE_URL}/storages/${uploadFilePath}`;
     }
 
-    fs.unlinkSync(formattedFilePath);
+    if (type === UploadType.IMAGE) fs.unlinkSync(formattedFilePath);
   }
 
+  fs.unlinkSync(filePath);
+
   return result;
+}
+
+export function getFilePathFromSeedData(sourceImageFileName: string) {
+  const assetPath = path.join(__dirname, `../../seed-data/assets`);
+
+  if (!existsSync(assetPath)) return '';
+
+  const files = fs.readdirSync(assetPath);
+  const file = files.find(e => {
+    const [imageFile] = e.split(/\.(jpeg|jpg|gif|png)$/);
+    if (imageFile === sourceImageFileName) return true;
+    return false;
+  });
+
+  if (!file) return '';
+
+  const tmpDir = os.tmpdir();
+  const extension = path.parse(file).ext;
+  const sourceImagePath = `${assetPath}/${file}`;
+  const targetImagePath = `${tmpDir}/${uuid()}${extension}`;
+
+  fs.copyFileSync(sourceImagePath, targetImagePath);
+
+  return targetImagePath;
+}
+
+function getMutations(type: UploadType): AnyObject {
+  if (type === UploadType.VIDEO) {
+    return {
+      format: 'mp4',
+      mutations: [
+        {
+          type: 'origin',
+          suffix: '',
+          width: 0,
+        },
+      ],
+    };
+  }
+
+  return {
+    format: 'jpg',
+    mutations: [
+      {
+        type: 'thumbnail',
+        suffix: '_thumbnail',
+        width: 200,
+      },
+      {
+        type: 'small',
+        suffix: '_small',
+        width: 400,
+      },
+      {
+        type: 'medium',
+        suffix: '_medium',
+        width: 600,
+      },
+      {
+        type: 'origin',
+        suffix: '',
+        width: 0,
+      },
+    ],
+  };
 }
